Extract building getCounts query as util function
